Move extension to a separate file
